Add files via upload
This Upper Confidence Bound Algorithm is a type of Reinforcement Learning algorithm that learns what advert to use given the data provided. The code goes through the first 10 people and looks at what they've chosen and then suggests its own advert to the next person (11th) and sees whether that person has clicked the advert and so on. It is learning at every stage.